requirements.txt
pandas>=1.3.3
openpyxl>=3.0.9
sqlalchemy>=1.4.25
numpy>=1.21.4
azure.data.tables
azure.storage.blob
pathlib>=1.0.1
python-dotenv>=1.0.1